+++ RELEASE-NOTES.md
@@ -14,6 +14,7 @@
 * #519 "FSFolder.include(..)" filter
 * #520 Zip source to exclude hidden files by default
 * #521 ZipEntry name extension filter for Zip
+* #522 "Zip.include(..)" filter
 
 ## Release 2.0.0-M3